Move processJSON function to the markdown_generators file and rename it generateMarkdownFromJson
