+++ src/lib/markdown_generators.js
@@ -9,6 +9,38 @@ const turndownService = new TurndownService()
 turndownService.use(tables)
 const { getCourseImageUrl } = require("./helpers")
 
+const generateMarkdownFromJson = courseData => {
+  /*
+    This function takes JSON data parsed from a master.json file and returns markdown data
+    */
+  const markdownData = []
+  let courseHomeMarkdown = generateCourseHomeFrontMatter(courseData)
+  courseHomeMarkdown += generateCourseFeatures(courseData)
+  courseHomeMarkdown += generateCourseCollections(courseData)
+  markdownData.push({
+    name: "_index.md",
+    data: courseHomeMarkdown
+  })
+  let menuIndex = 10
+  courseData["course_pages"].forEach(page => {
+    if (page["text"]) {
+      const pageName = page["short_url"]
+      let courseSectionMarkdown = generateCourseSectionFrontMatter(
+        page["title"],
+        menuIndex
+      )
+      courseSectionMarkdown += generateCourseSectionMarkdown(page, courseData)
+      const sectionData = {
+        name: `sections/${pageName}.md`,
+        data: courseSectionMarkdown
+      }
+      markdownData.push(sectionData)
+      menuIndex += 10
+    }
+  })
+  return markdownData
+}
+
 const generateCourseHomeFrontMatter = courseData => {
   /*
       Generate the front matter metadata for the course home page given course_data JSON
@@ -141,6 +173,7 @@ const generateCourseSectionMarkdown = (page, courseData) => {
 }
 
 module.exports = {
+  generateMarkdownFromJson,
   generateCourseHomeFrontMatter,
   generateCourseSectionFrontMatter,
   generateCourseFeatures,
